requirements.txt
influxdb-client
mysql-connector-python
